requirements.txt
django==3.2.12
django-admin-relation-links==0.2.5
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-model-utils==4.2.0
django-private-chat2==1.0.2
django-redis==5.2.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
channels==3.0.4
channels-redis==3.3.1
celery==5.2.3
googleapis-common-protos==1.53.0
grpcio==1.43.0
grpcio-tools==1.43.0
numpy==1.22.2
Pillow==7.0.0
python-decouple==3.5
requests==2.22.0
ring==0.9.1
robohash==1.1
scipy==1.8.0
gunicorn==20.1.0
psycopg2==2.9.3
SQLAlchemy==1.4.31
django-import-export==2.7.1
requests[socks]